WIP, replacing the backend with Runner
This particular bit I'm trying to replace the invocation loop with
runner where the idea is that runner will take its event output and
send it along to the jupyter interface.

There are some oddities and a lot of cleanup still needs to
happen. Right now this will only work with Runner installed from this
branch: